+++ lib/pure/complex.nim
@@ -25,8 +25,12 @@ type
   Complex* = tuple[re, im: float]
     ## a complex number, consisting of a real and an imaginary part
 
-proc toComplex*(x: SomeFloat): Complex =
-  ## Convert ``x`` to a complex number.
+const
+  im*: Complex = (re: 0.0, im: 1.0)
+    ## The imaginary unit. √-1.
+
+proc toComplex*(x: SomeInteger): Complex =
+  ## Convert some integer ``x`` to a complex number.
   result.re = x
   result.im = 0.0